@@ -2,13 +2,19 @@
   "name": "action-file-generator",
   "version": "1.0.0",
   "description": "The github action that creates index pages for your project",
-  "main": "out/action.js",
+  "main": "lib/action.js",
   "scripts": {
-    "build": "tsc && ncc build",
-    "format": "prettier --write '**/*.ts'",
-    "format-check": "prettier --check '**/*.ts'",
-    "lint": "eslint src/**/*.ts",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "compile": "tsc -b",
+    "build": "npm run clean && npm run compile",
+    "format": "prettier --write \"src/**/*.ts\" \"src/**/*.js\"",
+    "test": "mocha --debug-brk",
+    "clean": "rimraf lib",
+    "pretest": "npm run build",
+    "prepublishOnly": "npm test && npm run build",
+    "preversion": "",
+    "postversion": "git push && git push --tags",
+    "version": "git add -A src",
+    "watch": "tsc -w -p ./src"
   },
   "repository": {
     "type": "git",
@@ -17,8 +23,7 @@
   "keywords": [
     "Github",
     "action",
-    "markdown",
-    "index"
+    "file"
   ],
   "author": "DaanV2",
   "license": "ISC",
@@ -27,11 +32,18 @@
   },
   "homepage": "https://github.com/DaanV2/Action-File-Generator#readme",
   "devDependencies": {
-    "@types/node": "^14.14.34",
-    "tslint": "^5.12.1",
-    "typescript": "^3.3.3"
+    "@types/chai": "^4.2.21",
+    "@types/mocha": "^9.0.0",
+    "@types/node": "^16.6.1",
+    "chai": "^4.3.4",
+    "mocha": "^9.1.0",
+    "rimraf": "^3.0.2",
+    "ts-node": "^10.2.1",
+    "tslint": "^6.1.3",
+    "typescript": "^4.3.5"
   },
   "dependencies": {
-    "@actions/core": "^1.2.6"
+    "@actions/core": "^1.5.0",
+    "@actions/github": "^5.0.0"
   }
 }
