@@ -1,45 +1,42 @@
-import { CreateFolder } from './traverse';
-import * as fs from 'fs';
-
-const corexp = require('@actions/core');
+import * as fs from "fs";
+import * as core from "@actions/core";
+import { Process } from "./Types/process";
 
 //Start code
 try {
-	// This should be a token with access to your repository scoped in as a secret.
-	// The YML workflow will need to set myToken with the GitHub Secret Token
-	// token: ${{ secrets.GITHUB_TOKEN }}
-	const Folder = corexp.getInput('folder');
-	var result = false;
+  // This should be a token with access to your repository scoped in as a secret.
+  // The YML workflow will need to set myToken with the GitHub Secret Token
+  // token: ${{ secrets.GITHUB_TOKEN }}
+  const Folder = core.getInput("folder", { required: false });
+  const SpecFile = core.getInput("specification", { required: false });
 
-	console.log('starting on: ' + Folder);
+  var result = false;
 
-	if (fs.existsSync(Folder)) {
-		result = CreateFolder(Folder);
-	} else {
-		throw { message: 'Couldnt not find folder: ' + Folder };
-	}
+  console.log("starting on: " + SpecFile);
 
-	if (result) {
-		console.log('success');
-	}
-	else {
-		console.log('failure');
-		corexp.setFailed('no pages were created');
-	}
+  //processing
+  result = Process(SpecFile, Folder);
 
+  if (fs.existsSync(Folder)) {
+  } else {
+    throw { message: "Couldn't not find folder: " + Folder };
+  }
+
+  if (result) {
+    console.log("success");
+  } else {
+    console.log("failure");
+    core.setFailed("no pages were created");
+  }
 } catch (error) {
-	let message: string;
+  let message: string;
 
-	if (error.message)
-		message = error.message;
-	else
-		message = JSON.stringify(error);
+  if (error.message) message = error.message;
+  else message = JSON.stringify(error);
 
-	if (corexp)
-		corexp.setFailed(message);
-
-	else {
-		console.log(message);
-		process.exit(1);
-	}
+  if (core) core.setFailed(message);
+  else {
+    console.log(message);
+    process.exit(1);
+  }
 }
